part-4-patterns/starter/requirements.txt
strands-agents
strands-agents-tools
bedrock-agentcore[strands-agents]
strands-agents-tools[a2a_client]
strands-agents[a2a]
